refactor: group login controller dependencies in a single file
